+++ src/app/components/modal/modal.component.ts
@@ -1,30 +1,30 @@
 import { Component } from '@angular/core';
 import { FormsModule } from '@angular/forms';
-import { ModalProps } from '../../utils/types';
+import { ModalProps } from '../../utils/type';
 
 @Component({
   selector: 'app-modal',
   imports: [FormsModule],
   templateUrl: './modal.component.html',
-  styleUrl: './modal.component.css'
+  styleUrl: './modal.component.css',
 })
 export class ModalComponent {
   newGroup: ModalProps = {
-    nbGroups:0,
-    name: "",
+    nbGroups: 0,
+    name: '',
     isDWWM: false,
     age: false,
-    skill: false
+    skill: false,
+  };
+  createGroup(): void {
+    if (
+      this.newGroup.nbGroups !== null &&
+      this.newGroup.nbGroups !== undefined &&
+      this.newGroup.name.trim()
+    ) {
+      console.log(this.newGroup);
+    } else {
+      return alert('Veuillez remplir les 2 premiers champs');
+    }
   }
-createGroup(): void {
-  if (
-    this.newGroup.nbGroups !== null &&
-    this.newGroup.nbGroups !== undefined &&
-    this.newGroup.name.trim()
-  ) {
-    console.log(this.newGroup);
-  } else {
-    return alert('Veuillez remplir les 2 premiers champs');
-  }
-}
 }
