+++ src/app/utils/type.ts
@@ -20,3 +20,23 @@ export type ListProps = {
   techLevel: number,
   age: number
 }
+
+export type RegisterProps = {
+  firstName: string;
+  lastName: string;
+  email: string;
+  password: string;
+};
+
+export type LoginProps = {
+  email: string;
+  password: string;
+};
+
+export type ModalProps = {
+  nbGroups: number;
+  name: string;
+  isDWWM: boolean;
+  age: boolean;
+  skill: boolean;
+};
